Rename tampere parser to match api name
